code structure cleanup
