+++ src/lib/circuit.ts
@@ -1,5 +1,5 @@
 import { ProvablePure, Snarky } from '../snarky.js';
-import { MlFieldArray, MlFieldConstArray } from './field.js';
+import { MlFieldArray, MlFieldConstArray } from './ml/fields.js';
 import { withThreadPool } from '../bindings/js/wrapper.js';
 import { Provable } from './provable.js';
 import { snarkContext, gatesFromJson } from './provable-context.js';
