+++ src/lib/field.ts
@@ -10,16 +10,7 @@ import { MlArray } from './ml/base.js';
 export { Field };
 
 // internal API
-export {
-  ConstantField,
-  FieldType,
-  FieldVar,
-  FieldConst,
-  isField,
-  MlFieldArray,
-  MlFieldConstArray,
-  withMessage,
-};
+export { ConstantField, FieldType, FieldVar, FieldConst, isField, withMessage };
 
 const SnarkyFieldConstructor = SnarkyField(1).constructor;
 
@@ -1238,26 +1229,6 @@ const FieldBinable = defineBinable({
     ];
   },
 });
-
-type MlFieldArray = MlArray<FieldVar>;
-const MlFieldArray = {
-  to(arr: Field[]): MlArray<FieldVar> {
-    return MlArray.to(arr.map((x) => x.value));
-  },
-  from([, ...arr]: MlArray<FieldVar>) {
-    return arr.map((x) => new Field(x));
-  },
-};
-
-type MlFieldConstArray = MlArray<FieldConst>;
-const MlFieldConstArray = {
-  to(arr: Field[]): MlArray<FieldConst> {
-    return MlArray.to(arr.map((x) => x.toConstant().value[1]));
-  },
-  from([, ...arr]: MlArray<FieldConst>): ConstantField[] {
-    return arr.map((x) => new Field(x) as ConstantField);
-  },
-};
 
 function isField(x: unknown): x is Field {
   return x instanceof Field || (x as any) instanceof SnarkyFieldConstructor;
