+++ src/lib/hash-input.unit-test.ts
@@ -2,7 +2,6 @@ import {
   isReady,
   AccountUpdate,
   Types,
-  Field,
   Ledger,
   Permissions,
   shutdown,
@@ -16,10 +15,8 @@ import {
 } from '../bindings/mina-transaction/gen/transaction.js';
 import { packToFields } from './hash.js';
 import { Random, test } from './testing/property.js';
-import { OcamlInput } from '../snarky.js';
-import { MlArray } from './ml/base.js';
-import { MlFieldConstArray } from './field.js';
-import { Ml } from './ml/conversion.js';
+import { MlHashInput } from './ml/conversion.js';
+import { MlFieldConstArray } from './ml/fields.js';
 
 await isReady;
 
@@ -101,12 +98,12 @@ shutdown();
 
 function testInput<T, TJson>(
   Module: ProvableExtended<T, TJson>,
-  toInputOcaml: (json: string) => OcamlInput,
+  toInputOcaml: (json: string) => MlHashInput,
   value: T
 ) {
   let json = Module.toJSON(value);
   // console.log('json', json);
-  let input1 = inputFromOcaml(toInputOcaml(JSON.stringify(json)));
+  let input1 = MlHashInput.from(toInputOcaml(JSON.stringify(json)));
   let input2 = Module.toInput(value);
   // console.log('snarkyjs', JSON.stringify(input2));
   // console.log();
@@ -115,7 +112,7 @@ function testInput<T, TJson>(
   expect(JSON.stringify(input2)).toEqual(JSON.stringify(input1));
   // console.log('ok?', ok1);
   let fields1 = MlFieldConstArray.from(
-    Ledger.hashInputFromJson.packInput(inputToOcaml(input1))
+    Ledger.hashInputFromJson.packInput(MlHashInput.to(input1))
   );
   let fields2 = packToFields(input2);
   let ok2 = JSON.stringify(fields1) === JSON.stringify(fields2);
@@ -124,30 +121,6 @@ function testInput<T, TJson>(
   if (!ok1 || !ok2) {
     throw Error('inconsistent toInput');
   }
-}
-
-function inputFromOcaml([, fields, packed]: OcamlInput) {
-  return {
-    fields: MlFieldConstArray.from(fields),
-    packed: MlArray.from(packed).map(
-      ([, field, size]) => [Field(field), size] as [Field, number]
-    ),
-  };
-}
-function inputToOcaml({
-  fields,
-  packed,
-}: {
-  fields: Field[];
-  packed: [field: Field, size: number][];
-}): OcamlInput {
-  return [
-    0,
-    MlFieldConstArray.to(fields),
-    MlArray.to(
-      packed.map(([field, size]) => [0, Ml.constFromField(field), size])
-    ),
-  ];
 }
 
 function fixVerificationKey(accountUpdate: Json.AccountUpdate) {
